+++ merk/src/merk/defaults.rs
@@ -34,4 +34,4 @@ pub const ROOT_KEY_KEY: &[u8] = b"r";
 #[cfg(feature = "full")]
 pub const MAX_UPDATE_VALUE_BASED_ON_COSTS_TIMES: u8 = 8;
 #[cfg(feature = "full")]
-pub const MAX_PREFIXED_KEY_SIZE: u32 = 288;
+pub const MAX_PREFIXED_KEY_SIZE: u64 = 288;
