+++ storage/src/rocksdb_storage/storage_context/context_tx.rs
@@ -240,7 +240,7 @@ impl<'db> StorageContext<'db> for PrefixedRocksDbTransactionContext<'db> {
                     .as_ref()
                     .ok()
                     .and_then(Option::as_ref)
-                    .map(|x| x.len() as u32)
+                    .map(|x| x.len() as u64)
                     .unwrap_or(0),
                 ..Default::default()
             })
@@ -256,7 +256,7 @@ impl<'db> StorageContext<'db> for PrefixedRocksDbTransactionContext<'db> {
                     .as_ref()
                     .ok()
                     .and_then(Option::as_ref)
-                    .map(|x| x.len() as u32)
+                    .map(|x| x.len() as u64)
                     .unwrap_or(0),
                 ..Default::default()
             })
@@ -272,7 +272,7 @@ impl<'db> StorageContext<'db> for PrefixedRocksDbTransactionContext<'db> {
                     .as_ref()
                     .ok()
                     .and_then(Option::as_ref)
-                    .map(|x| x.len() as u32)
+                    .map(|x| x.len() as u64)
                     .unwrap_or(0),
                 ..Default::default()
             })
@@ -288,7 +288,7 @@ impl<'db> StorageContext<'db> for PrefixedRocksDbTransactionContext<'db> {
                     .as_ref()
                     .ok()
                     .and_then(Option::as_ref)
-                    .map(|x| x.len() as u32)
+                    .map(|x| x.len() as u64)
                     .unwrap_or(0),
                 ..Default::default()
             })
